+++ Assets/CommonCoreModules/World/ObjectComponents/IControlPlayerCamera.cs
@@ -7,5 +7,6 @@ namespace CommonCore.World
     public interface IControlPlayerCamera
     {
         Camera GetCamera();
+        AudioListener GetAudioListener();
     }
 }