+++ CHANGELOG.md
@@ -374,7 +374,9 @@
 * ActorController now catches exceptions thrown by OnDeathSpecial
 * GetParamsForModule now checks if there's actually a delimiter after the module name
 * GetActiveCamera getting player camera is now delegated to PlayerController via IControlPlayerCamera interface
+* IControlPlayerCamera now provides getter for player audio listener
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]
 * Added NavigationNodeReachedTrigger action trigger [Untested]
+* Added audio listener switching methods to WorldUtils [Untested]